feat: add mock chain and state
