+++ vm/src/tests/mod.rs
@@ -1,2 +1,4 @@
 pub mod simple_transaction_tests;
 pub mod vm_error_tests;
+
+mod mock_chain;
