+++ vm/src/tests/simple_transaction_tests.rs
@@ -1,18 +1,14 @@
 use crate::{
-    access_path::AccessPath,
     gas::Gas,
     message::{EntryFunction, Message, Module, ModuleBundle, Script},
     nova_vm::NovaVM,
     storage::data_view_resolver::DataViewResolver,
-    storage::state_view::StateView,
 };
-use std::collections::BTreeMap;
 
 use move_deps::{
     move_binary_format::CompiledModule,
     move_core_types::{
         account_address::AccountAddress,
-        effects::{ChangeSet, Op},
         identifier::Identifier,
         language_storage::{ModuleId, TypeTag},
         parser::parse_struct_tag,
@@ -23,55 +19,6 @@ use move_deps::{
 use crate::asset::{
     compile_move_nursery_modules, compile_move_stdlib_modules, compile_nova_stdlib_modules,
 };
-
-//faking chain db
-struct MockDB {
-    map: BTreeMap<AccessPath, Option<Vec<u8>>>,
-}
-
-impl MockDB {
-    pub fn new() -> Self {
-        Self {
-            map: BTreeMap::new(),
-        }
-    }
-
-    fn write_op(&mut self, ref ap: AccessPath, ref blob_opt: Op<Vec<u8>>) {
-        match blob_opt {
-            Op::New(blob) | Op::Modify(blob) => {
-                self.map.insert(ap.clone(), Some(blob.clone()));
-            }
-            Op::Delete => {
-                self.map.remove(ap);
-                self.map.insert(ap.clone(), None);
-            }
-        }
-    }
-
-    pub fn push_write_set(&mut self, changeset: ChangeSet) {
-        for (addr, account_changeset) in changeset.into_inner() {
-            let (modules, resources) = account_changeset.into_inner();
-            for (struct_tag, blob_opt) in resources {
-                let ap = AccessPath::resource_access_path(addr, struct_tag);
-                self.write_op(ap, blob_opt)
-            }
-
-            for (name, blob_opt) in modules {
-                let ap = AccessPath::from(&ModuleId::new(addr, name));
-                self.write_op(ap, blob_opt)
-            }
-        }
-    }
-}
-
-impl StateView for MockDB {
-    fn get(&self, access_path: &AccessPath) -> anyhow::Result<Option<Vec<u8>>> {
-        match self.map.get(access_path) {
-            Some(opt_data) => Ok(opt_data.clone()),
-            None => Ok(None),
-        }
-    }
-}
 
 #[cfg(test)]
 impl Module {
@@ -206,7 +153,9 @@ impl ExpectedOutput {
 }
 
 fn run_transaction(testcases: Vec<(Message, ExpectedOutput)>) {
-    let mut db = MockDB::new();
+    use super::mock_chain::MockChain;
+
+    let mut chain = MockChain::new();
     let mut vm = NovaVM::new();
 
     // publish move_stdlib and move_nursery and nova_stdlib modules
@@ -214,8 +163,9 @@ fn run_transaction(testcases: Vec<(Message, ExpectedOutput)>) {
     modules.append(&mut compile_move_nursery_modules());
     modules.append(&mut compile_nova_stdlib_modules());
 
+    let mut state = chain.create_state();
     for module in modules {
-        let resolver = DataViewResolver::new(&db);
+        let resolver = DataViewResolver::new(&state);
         let mut mod_blob = vec![];
         module
             .serialize(&mut mod_blob)
@@ -224,12 +174,14 @@ fn run_transaction(testcases: Vec<(Message, ExpectedOutput)>) {
             .initialize(mod_blob, &resolver)
             .expect("Module must load");
         assert!(status == VMStatus::Executed);
-        db.push_write_set(output.change_set().clone());
+        state.push_write_set(output.change_set().clone());
     }
+    chain.commit(state);
 
     let gas_limit = Gas::new(100_000u64);
     for (msg, exp_output) in testcases {
-        let resolver = DataViewResolver::new(&db);
+        let mut state = chain.create_state();
+        let resolver = DataViewResolver::new(&state);
         let (status, output, result) = vm
             .execute_message(msg, &resolver, gas_limit)
             .expect("nova vm failure");
@@ -245,7 +197,9 @@ fn run_transaction(testcases: Vec<(Message, ExpectedOutput)>) {
             continue;
         }
         // apply output into db
-        db.push_write_set(output.change_set().clone());
+        state.push_write_set(output.change_set().clone());
+
+        chain.commit(state);
 
         // TODO: mock tx rollback
         vm.invalidate_loader_cache();
